+++ varinfo/exceptions.py
@@ -65,7 +65,7 @@ class MissingConfigurationFileError(CustomError):
 
 
 class CMRQueryException(CustomError):
-    ''' This exception is raised when a query to CMR fails. 
+    ''' This exception is raised when a query to CMR fails.
     '''
 
     def __init__(self, cmr_exception_message):
@@ -91,3 +91,12 @@ class MissingGranuleDownloadLinks(CustomError):
     def __init__(self, download_link):
         super().__init__('MissingGranuleDownloadLinks',
                          f'No links for granule record: {download_link}')
+
+
+class RequestsException(CustomError):
+    ''' This exception is raised when a query to CMR fails.
+    '''
+
+    def __init__(self, requests_exception_message):
+        super().__init__('RequestsException',
+                         f'requests module failed with the following error: {requests_exception_message}')